muwon

+++ final/src/_gcc_phat.cpp
@@ -1,10 +1,10 @@
-#include "CalDegree.hpp"
+#include "../include/CalDegree.hpp"
 using namespace std;
 using namespace CalDegree;
 
 namespace CalDegree {
     // GCC-PHAT 계산 함수
-    int _gcc_phat(const vector<double> &x, const vector<double> &y) {
+    double _gcc_phat(const vector<double> &x, const vector<double> &y) {
         int N = x.size();
 
         // 입력 신호를 복소수로 변환 (정상적인 FFT를 위해)
@@ -38,7 +38,7 @@ namespace CalDegree {
         }
 
         // 시간 지연 (초 단위로 반환)
-        int delay_samples = max_index;
+        double delay_samples = max_index;
         if (delay_samples > N / 2) {
             delay_samples -= N;  // 시간 지연이 음수일 수 있기 때문
         }
